+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.1.10
 - onPickupCollected callback : check for non item pickup
+- benthicBloom : now using Item.StackKind.NORMAL every time
 
 ### v0.1.9
 - PlasmaShrimp and Polylute : added checks for instance before using fire_direct